Simple application with user registration and auth, upload, view, etc.

+++ garvis-backend/app/app.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI, HTTPException, File, UploadFile, Form, Depends
 from app.api.health import router as health_router
 from app.database.duckdb_data_service import DataService
-from app.schemas.post_schemas import PostCreate, PostResponse
+from app.schemas.post_schemas import PostCreate, PostResponse, UserRead, UserCreate, UserUpdate
 from app.database.sqlite_data_service import Post, create_db_and_tables, get_async_session
 from sqlalchemy.ext.asyncio import AsyncSession
 from contextlib import asynccontextmanager
@@ -12,6 +12,8 @@ import shutil
 import os
 import uuid
 import tempfile
+from app.users import auth_backend, current_active_user, fastapi_users
+from app.database.sqlite_data_service import User
 
 @asynccontextmanager
 async def lifespan(app: FastAPI):
@@ -30,11 +32,19 @@ app = FastAPI(title="Garvis Backend", version="0.1.0",lifespan=lifespan)
 #    return {"name": "garvis-backend", "status": "ok"}
 ######################################################################
 
+#include all endpoints in the "fastapi_users.get_auth_router" router and prefix them with "/auth/jwt"
+app.include_router(fastapi_users.get_auth_router(auth_backend), prefix="/auth/jwt", tags=["auth"])
+app.include_router(fastapi_users.get_register_router(UserRead, UserCreate), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_reset_password_router(), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_verify_router(UserRead), prefix="/auth", tags=["auth"])
+app.include_router(fastapi_users.get_users_router(UserRead,UserUpdate), prefix="/users", tags=["users"])
+
 @app.post("/upload")
 async def upload_file(
     file: UploadFile=File(...)
     , caption:str=Form("")
     , session:AsyncSession = Depends(get_async_session)
+    , user: User = Depends(current_active_user)
 ):
     temp_file_path = None
 
@@ -55,9 +65,10 @@ async def upload_file(
         if(upload_result.response_metadata.http_status_code==200):
 
             post = Post(
-                caption=caption
+                user_id = user.id
+                , caption=caption
                 , url=upload_result.url
-                , file_type = "video" if file.content_type.startswith("video/") else "iamge"
+                , file_type = "video" if file.content_type.startswith("video/") else "image"
                 , file_name = upload_result.name
             )
 
@@ -76,7 +87,9 @@ async def upload_file(
 
 
 @app.delete("/posts/{post_id}")
-async def delete_post(post_id:str, session: AsyncSession = Depends(get_async_session)):
+async def delete_post(post_id:str
+                        , session: AsyncSession = Depends(get_async_session)
+                        , user: User = Depends(current_active_user)):
     try:
         post_uuid = uuid.UUID(post_id)
         result = await session.execute(Select(Post).where(Post.id == post_uuid))
@@ -84,6 +97,9 @@ async def delete_post(post_id:str, session: AsyncSession = Depends(get_async_ses
 
         if(not post):
             raise HTTPException(status_code=404, detail="Post not found")
+        
+        if(post.user_id != user.id):
+            raise HTTPException(status_code=403, detail="You dont have permission to delete this post")
         
         await session.delete(post)
         await session.commit()
@@ -94,9 +110,14 @@ async def delete_post(post_id:str, session: AsyncSession = Depends(get_async_ses
 
 
 @app.get("/feed")
-async def get_feed(session:AsyncSession = Depends(get_async_session)):
+async def get_feed(session:AsyncSession = Depends(get_async_session)
+                    , user: User = Depends(current_active_user)):
     result = await session.execute(Select(Post).order_by(Post.created_at.desc()))
     posts = [row[0] for row in result.all()]
+
+    result = await session.execute(Select(User))
+    users = [row[0] for row in result.all()]
+    user_dict = {u.id:u.email for u in users}
 
     posts_data = []
 
@@ -104,11 +125,14 @@ async def get_feed(session:AsyncSession = Depends(get_async_session)):
         posts_data.append(
             {
                 "id": str(post.id)
-                , "caption": post.caption
-                , "url": post.url
-                , "file_type": post.file_type
-                , "file_name": post.file_name
-                , "create_at": post.created_at.isoformat()
+                ,"user_id":str(post.user_id)
+                ,"caption": post.caption
+                ,"url": post.url
+                ,"file_type": post.file_type
+                ,"file_name": post.file_name
+                ,"created_at": post.created_at.isoformat()
+                ,"is_owner": post.user_id == user.id
+                ,"email":user_dict.get(post.user_id,"Unknown User")
             }
         )
